# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "DashboardCreate": {
        "properties": {
          "dataset_id": {
            "title": "Dataset Id",
            "type": "string"
          },
          "description": {
            "default": "",
            "maxLength": 500,
            "title": "Description",
            "type": "string"
          },
          "name": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "spec": {
            "$ref": "#/components/schemas/DashboardSpec"
          }
        },
        "required": [
          "name",
          "dataset_id",
          "spec"
        ],
        "title": "DashboardCreate",
        "type": "object"
      },
      "DashboardOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "dataset_id": {
            "title": "Dataset Id",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "spec": {
            "$ref": "#/components/schemas/DashboardSpec"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "dataset_id",
          "spec",
          "created_at",
          "updated_at"
        ],
        "title": "DashboardOut",
        "type": "object"
      },
      "DashboardRunOut": {
        "properties": {
          "dashboard": {
            "$ref": "#/components/schemas/DashboardOut"
          },
          "result": {
            "$ref": "#/components/schemas/DatasetQueryResult"
          }
        },
        "required": [
          "dashboard",
          "result"
        ],
        "title": "DashboardRunOut",
        "type": "object"
      },
      "DashboardSpec": {
        "properties": {
          "query": {
            "$ref": "#/components/schemas/DatasetQuery"
          },
          "visualization": {
            "default": "table",
            "enum": [
              "table",
              "bar",
              "line",
              "donut"
            ],
            "title": "Visualization",
            "type": "string"
          },
          "x_field": {
            "anyOf": [
              {
                "maxLength": 160,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "X Field"
          },
          "y_fields": {
            "items": {
              "type": "string"
            },
            "maxItems": 10,
            "title": "Y Fields",
            "type": "array"
          }
        },
        "title": "DashboardSpec",
        "type": "object"
      },
      "DatasetColumn": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "nullable": {
            "title": "Nullable",
            "type": "boolean"
          },
          "type": {
            "enum": [
              "boolean",
              "integer",
              "number",
              "string",
              "date",
              "datetime"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "name",
          "type",
          "nullable"
        ],
        "title": "DatasetColumn",
        "type": "object"
      },
      "DatasetCreate": {
        "properties": {
          "description": {
            "default": "",
            "maxLength": 500,
            "title": "Description",
            "type": "string"
          },
          "name": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          }
        },
        "required": [
          "name",
          "source_id"
        ],
        "title": "DatasetCreate",
        "type": "object"
      },
...
      "DatasetOut": {
        "properties": {
          "columns": {
            "items": {
              "$ref": "#/components/schemas/DatasetColumn"
            },
            "title": "Columns",
            "type": "array"
          },
          "content_hash": {
            "title": "Content Hash",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "current_version": {
            "title": "Current Version",
            "type": "integer"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "format": {
            "title": "Format",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "row_count": {
            "title": "Row Count",
            "type": "integer"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version_id": {
            "title": "Version Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "current_version",
          "version_id",
          "source_id",
          "format",
          "columns",
          "row_count",
          "content_hash",
          "created_at",
          "updated_at"
        ],
        "title": "DatasetOut",
        "type": "object"
      },
      "DatasetQuery": {
        "properties": {
          "filters": {
            "items": {
              "$ref": "#/components/schemas/DatasetFilter"
            },
            "maxItems": 20,
            "title": "Filters",
            "type": "array"
          },
          "group_by": {
            "anyOf": [
              {
                "maxLength": 160,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group By"
          },
          "limit": {
            "default": 100,
            "maximum": 500.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "metrics": {
            "items": {
              "$ref": "#/components/schemas/DatasetMetric"
            },
            "maxItems": 10,
            "title": "Metrics",
            "type": "array"
          },
          "order_by": {
            "anyOf": [
              {
                "maxLength": 160,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order By"
          },
          "order_direction": {
            "default": "asc",
            "enum": [
              "asc",
              "desc"
            ],
            "title": "Order Direction",
            "type": "string"
          }
        },
        "title": "DatasetQuery",
        "type": "object"
      },
      "DatasetQueryResult": {
        "properties": {
          "columns": {
            "items": {
              "type": "string"
            },
            "title": "Columns",
            "type": "array"
          },
          "elapsed_ms": {
            "title": "Elapsed Ms",
            "type": "number"
          },
          "row_count": {
            "title": "Row Count",
            "type": "integer"
          },
          "rows": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Rows",
            "type": "array"
          },
          "truncated": {
            "title": "Truncated",
            "type": "boolean"
          }
        },
        "required": [
          "columns",
          "rows",
          "row_count",
          "truncated",
          "elapsed_ms"
        ],
        "title": "DatasetQueryResult",
        "type": "object"
      },
      "DatasetVersionCreate": {
        "properties": {
          "source_id": {
            "title": "Source Id",
            "type": "string"
          }
        },
        "required": [
          "source_id"
        ],
        "title": "DatasetVersionCreate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Durable chat, cited local knowledge, provenance, and permissioned read-only tools.",
    "title": "Agentic Knowledge Workspace API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/dashboards": {
      "get": {
        "operationId": "list_dashboards_api_dashboards_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-user-id",
            "required": false,
            "schema": {
              "default": "00000000-0000-4000-8000-000000000002",
              "title": "X-User-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-workspace-id",
            "required": false,
            "schema": {
              "default": "00000000-0000-4000-8000-000000000001",
              "title": "X-Workspace-Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DashboardOut"
                  },
                  "title": "Response List Dashboards Api Dashboards Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Dashboards",
        "tags": [
          "analytics"
        ]
      },
      "post": {
        "operationId": "create_dashboard_api_dashboards_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": false,
            "schema": {
              "default": "00000000-0000-4000-8000-000000000002",
              "title": "X-User-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-workspace-id",
            "required": false,
            "schema": {
              "default": "00000000-0000-4000-8000-000000000001",
              "title": "X-Workspace-Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DashboardCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Dashboard",
        "tags": [
          "analytics"
        ]
      }
    },
    "/api/dashboards/{dashboard_id}/run": {
      "post": {
        "operationId": "run_dashboard_api_dashboards__dashboard_id__run_post",
        "parameters": [
          {
            "in": "path",
            "name": "dashboard_id",
            "required": true,
            "schema": {
              "title": "Dashboard Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": false,
            "schema": {
              "default": "00000000-0000-4000-8000-000000000002",
              "title": "X-User-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-workspace-id",
            "required": false,
            "schema": {
              "default": "00000000-0000-4000-8000-000000000001",
              "title": "X-Workspace-Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardRunOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Dashboard",
        "tags": [
          "analytics"
        ]
      }
    },
    "/api/datasets": {
      "get": {
        "operationId": "list_datasets_api_datasets_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-user-id",
            "required": false,
            "schema": {
              "default": "00000000-0000-4000-8000-000000000002",
              "title": "X-User-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-workspace-id",
            "required": false,
            "schema": {
              "default": "00000000-0000-4000-8000-000000000001",
              "title": "X-Workspace-Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DatasetOut"
                  },
                  "title": "Response List Datasets Api Datasets Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Datasets",
        "tags": [
          "analytics"
        ]
      },
      "post": {
        "operationId": "create_dataset_api_datasets_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": false,
            "schema": {
              "default": "00000000-0000-4000-8000-000000000002",
              "title": "X-User-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-workspace-id",
            "required": false,
            "schema": {
              "default": "00000000-0000-4000-8000-000000000001",
              "title": "X-Workspace-Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatasetCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Dataset",
        "tags": [
          "analytics"
        ]
      }
    },
    "/api/datasets/{dataset_id}/query": {
      "post": {
        "operationId": "query_dataset_api_datasets__dataset_id__query_post",
        "parameters": [
          {
            "in": "path",
            "name": "dataset_id",
            "required": true,
            "schema": {
              "title": "Dataset Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": false,
            "schema": {
              "default": "00000000-0000-4000-8000-000000000002",
              "title": "X-User-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-workspace-id",
            "required": false,
            "schema": {
              "default": "00000000-0000-4000-8000-000000000001",
              "title": "X-Workspace-Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatasetQuery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetQueryResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Query Dataset",
        "tags": [
          "analytics"
        ]
      }
    },
    "/api/datasets/{dataset_id}/versions": {
      "post": {
        "operationId": "create_version_api_datasets__dataset_id__versions_post",
        "parameters": [
          {
            "in": "path",
            "name": "dataset_id",
            "required": true,
            "schema": {
              "title": "Dataset Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": false,
            "schema": {
              "default": "00000000-0000-4000-8000-000000000002",
              "title": "X-User-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-workspace-id",
            "required": false,
            "schema": {
              "default": "00000000-0000-4000-8000-000000000001",
              "title": "X-Workspace-Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatasetVersionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Version",
        "tags": [
          "analytics"
        ]
      }
    },
...
  }
}